+++ ctrl-lib/translator/PolarCoordTranslator.cpp
@@ -40,9 +40,9 @@ namespace CalX {
 		this->base = t;
 	}
 	
-	motor_point_t PolarCoordTranslator::get(long double x, long double y) {
-		long double nx = x * cos(y);
-		long double ny = x * sin(y);
+	motor_point_t PolarCoordTranslator::get(double x, double y) {
+		double nx = x * cos(y);
+		double ny = x * sin(y);
 		if (this->base == nullptr) {
 			motor_point_t pnt = {(int64_t) nx, (int64_t) ny};
 			return pnt;
@@ -54,12 +54,12 @@ namespace CalX {
 	coord_point_t PolarCoordTranslator::get(motor_point_t pnt) {
 		coord_point_t out;
 		if (this->base == nullptr) {
-			out = {(long double) pnt.x, (long double) pnt.y};
+			out = {(double) pnt.x, (double) pnt.y};
 		} else {
 			out = this->base->get(pnt);
 		}
-		long double p = sqrt(pow(out.x, 2) + pow(out.y, 2));
-		long double f = atan2(out.y, out.x);
+		double p = sqrt(pow(out.x, 2) + pow(out.y, 2));
+		double f = atan2(out.y, out.x);
 		out = {p, f};
 		return out;
 	}
